pipeline/pyproject.toml
[project]
name = "pipeline"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.22",
    "dagster-duckdb>=0.28.22",
    "dagster-duckdb-pandas>=0.28.22",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "pipeline"
registry_modules = [
    "pipeline.components.*",
]


